refactor: move E2E tests from playground/testRun.ts into separate files
