+++ test/playground/pages/navigate-early/navigateEarly.e2e-test.ts
@@ -1,0 +1,16 @@
+import { autoRetry, getServerUrl, page, test } from '@brillout/test-e2e'
+import { expectUrl } from '../../../utils'
+
+function testNavigateEarly() {
+  test('Calling navigate() early in +client.js', async () => {
+    await page.goto(getServerUrl() + '/navigate-early')
+    await autoRetry(
+      () => {
+        expectUrl('/markdown')
+      },
+      { timeout: 5000 }
+    )
+  })
+}
+
+export default [testNavigateEarly]
