+++ test/playground/pages/redirects.e2e-tests.ts
@@ -1,0 +1,10 @@
+import { test, getServerUrl, expect } from '@brillout/test-e2e'
+
+function testRedirectMailto() {
+  test('Redirect to URI without http protocol (e.g. `mailto:`)', async () => {
+    const resp = await fetch(getServerUrl() + '/mail', { redirect: 'manual' })
+    expect(resp.headers.get('Location')).toBe('mailto:some@example.com')
+  })
+}
+
+export default [testRedirectMailto]
